renamed to sonde_like data since we can read IAGOS data in HARP format with this reader as well
